requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
docutils==0.20.1
ffpyplayer==4.5.0
greenlet==2.0.2
idna==3.4
Kivy==2.2.1
Kivy-Garden==0.1.5
Pillow==9.5.0
psycopg2-binary==2.9.6
Pygments==2.15.1
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
typing_extensions==4.6.3
urllib3==2.0.3
